refactor styles for status
