+++ src/components/layouts/vales/UpdateVale.tsx
@@ -18,14 +18,7 @@ import { useStatusStore } from "@/stores/valesStore";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { useToast } from "@/hooks/use-toast";
 import { useUpdateVales } from "@/hooks/Vales/useUpdateVale";
-
-const statusStyles = {
-  pendiente: "bg-amber-300 text-amber-950",
-  progreso: "bg-sky-300 text-sky-950",
-  completada: "bg-green-300 text-green-950 ",
-  cancelada: "bg-rose-300 text-red-950 ",
-  incompleto: "bg-stone-300 text-stone-950 ",
-} as const;
+import statusStyles from "./statusStyles";
 
 export function UpdateVale() {
   const {
