+++ src/components/layouts/vales/profesor/ValeDetailsProfesor.tsx
@@ -10,14 +10,7 @@ import { Badge } from "@/components/ui/badge";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { Separator } from "@/components/ui/separator";
 import { ValeProfesorDetails } from "@/types/ValeTypes";
-
-const statusStyles = {
-  pendiente: "bg-amber-300 text-amber-950",
-  progreso: "bg-sky-300 text-sky-950",
-  completada: "bg-green-300 text-green-950",
-  cancelada: "bg-rose-300 text-red-950",
-  incompleto: "bg-stone-300 text-stone-950",
-} as const;
+import statusStyles from "../statusStyles";
 
 interface PracticaAsignadaDisplayProps {
   data: ValeProfesorDetails;
